Use the test watcher in all unit tests
