+++ src/test/java/HttpEventCollector_LogbackTest.java
@@ -23,7 +23,9 @@ import com.splunk.logging.HttpEventCollectorEventInfo;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Assert;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TestRule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,6 +34,9 @@ public final class HttpEventCollector_LogbackTest {
     private String httpEventCollectorName = "LogbackTest";
     List<List<HttpEventCollectorEventInfo>> errors = new ArrayList<List<HttpEventCollectorEventInfo>>();
     List<HttpEventCollectorErrorHandler.ServerErrorException> logEx = new ArrayList<HttpEventCollectorErrorHandler.ServerErrorException>();
+
+    @Rule
+    public TestRule watcher = new SplunkTestWatcher();
 
     /**
      * sending a message via httplogging using logback to splunk
@@ -361,7 +366,6 @@ public final class HttpEventCollector_LogbackTest {
         TestUtil.verifyEventsSentInOrder(prefix, totalEventsCount, indexName);
 
         TestUtil.deleteHttpEventCollectorToken(httpEventCollectorName);
-        System.out.println("====================== Test pass=========================");
     }
     
     /**
